backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
celery==5.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
pysnmp==6.1.4
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
email-validator==2.1.0.post1
httpx==0.26.0
